% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_dataset}
\alias{get_dataset}
\title{Get Dataset}
\usage{
get_dataset(id, query = "SELECT *")
}
\arguments{
\item{id}{a string}

\item{query}{(optional) an SoQL query ("Socrata Query Language")}
}
\value{
a dataset
}
\description{
Look for a dataset here:
}
\details{
https://www.datos.gov.co/browse?sortBy=most_accessed&pageSize=50&category=Justicia+y+Derecho&provenance=official
}
\examples{
\dontrun{
divipola <- get_dataset("gdxc-w37w")

d <- get_dataset(
  id = "g3ma-7zce",
  query = "SELECT * WHERE anio = '2024'"
)

}

}
